# server/docs/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "DMetrics Backend REST API",
    "version": "1.0.0",
    "description": "Production-grade REST API backend powering Users, Projects, and Tasks for the DMetrics Developer Productivity & Engineering Velocity Platform.",
    "contact": {
      "name": "DMetrics Engineering Team",
      "email": "engineering@dmetrics.dev"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Local Development Server"
    }
  ],
...
  "paths": {
...
    "/projects/{id}": {
      "get": {
        "tags": [
          "Projects"
        ],
        "summary": "Get project by ID or key",
        "description": "Finds a project by its unique ID (e.g. `proj_1`) or project key (e.g. `CPE`).",
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Project found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Projects"
        ],
        "summary": "Full update of project",
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          },
          "409": {
            "$ref": "#/components/responses/409Conflict"
          }
        }
      },
      "patch": {
        "tags": [
          "Projects"
        ],
        "summary": "Partial update of project",
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Project updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Projects"
        ],
        "summary": "Delete project",
        "description": "Deletes project and cascades deletion to associated project tasks. Requires Staff, Lead, or Architect privileges.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "responses": {
          "204": {
            "description": "Project deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/401Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/403Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          }
        }
      }
    },
...
    "/tasks/{id}": {
      "get": {
        "tags": [
          "Tasks"
        ],
        "summary": "Get task by ID or key",
        "description": "Retrieve task by task ID (e.g. `task_1`) or task key (e.g. `CPE-104`).",
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "responses": {
          "200": {
            "description": "Task found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          }
        }
      },
      "put": {
        "tags": [
          "Tasks"
        ],
        "summary": "Full update of task",
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          }
        }
      },
      "patch": {
        "tags": [
          "Tasks"
        ],
        "summary": "Partial update of task",
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Task updated",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/400BadRequest"
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          }
        }
      },
      "delete": {
        "tags": [
          "Tasks"
        ],
        "summary": "Delete task",
        "description": "Deletes a task by ID or key. Requires authentication.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "$ref": "#/components/parameters/idParam"
          }
        ],
        "responses": {
          "204": {
            "description": "Task deleted successfully"
          },
          "401": {
            "$ref": "#/components/responses/401Unauthorized"
          },
          "404": {
            "$ref": "#/components/responses/404NotFound"
          }
        }
      }
    },
...
  },
  "components": {
    "parameters": {
      "idParam": {
        "name": "id",
        "in": "path",
        "required": true,
        "description": "Unique identifier or alphanumeric entity key",
        "schema": {
          "type": "string"
        }
      }
    },
    "responses": {
      "400BadRequest": {
        "description": "Validation failure or malformed request payload",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "example": {
              "success": false,
              "statusCode": 400,
              "error": "ValidationError",
              "message": "Validation failed for request data",
              "errors": [
                {
                  "field": "status",
                  "message": "Status must be one of: 'backlog', 'in_progress', 'in_review', 'done'"
                }
              ],
              "timestamp": "2026-09-06T07:00:00.000Z",
              "path": "/api/tasks/task_1/status"
            }
          }
        }
      },
      "404NotFound": {
        "description": "Requested resource does not exist",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "example": {
              "success": false,
              "statusCode": 404,
              "error": "ApiError",
              "message": "Task with ID 'task_999' not found",
              "timestamp": "2026-09-06T07:00:00.000Z",
              "path": "/api/tasks/task_999"
            }
          }
        }
      },
      "409Conflict": {
        "description": "Unique constraint violation (e.g. duplicate key or email)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "example": {
              "success": false,
              "statusCode": 409,
              "error": "ConflictError",
              "message": "User with email 'alex.chen@innovate.dev' already exists",
              "timestamp": "2026-09-06T07:00:00.000Z",
              "path": "/api/users"
            }
          }
        }
      },
      "401Unauthorized": {
        "description": "Unauthorized: Bearer token is missing, invalid, or expired",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "example": {
              "success": false,
              "statusCode": 401,
              "error": "UnauthorizedError",
              "message": "Authentication required: Access token is missing or malformed",
              "timestamp": "2026-09-06T07:00:00.000Z",
              "path": "/api/auth/me"
            }
          }
        }
      },
      "403Forbidden": {
        "description": "Forbidden: User role lacks sufficient engineering privileges (RBAC)",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ApiErrorResponse"
            },
            "example": {
              "success": false,
              "statusCode": 403,
              "error": "ForbiddenError",
              "message": "Forbidden: Role 'Junior Frontend Developer' lacks sufficient privileges. Required roles: Staff, Lead, Admin, Architect, Principal",
              "timestamp": "2026-09-06T07:00:00.000Z",
              "path": "/api/auth/verify-lead"
            }
          }
        }
      }
    },
    "schemas": {
      "ApiErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "statusCode": {
            "type": "integer",
            "example": 400
          },
          "error": {
            "type": "string",
            "example": "ValidationError"
          },
          "message": {
            "type": "string",
            "example": "Validation failed for request data"
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "field": {
                  "type": "string"
                },
                "message": {
                  "type": "string"
                }
              }
            }
          },
          "timestamp": {
            "type": "string"
          },
          "path": {
            "type": "string"
          }
        }
      },
      "User": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "usr_1"
          },
          "name": {
            "type": "string",
            "example": "Alex Chen"
          },
          "email": {
            "type": "string",
            "example": "alex.chen@innovate.dev"
          },
          "role": {
            "type": "string",
            "example": "Staff Platform Engineer"
          },
          "username": {
            "type": "string",
            "example": "alexchen-dev"
          },
          "avatar": {
            "type": "string"
          },
          "bio": {
            "type": "string"
          },
          "location": {
            "type": "string"
          },
          "timezone": {
            "type": "string"
          },
          "githubUsername": {
            "type": "string"
          },
          "githubUrl": {
            "type": "string"
          },
          "productivityScore": {
            "type": "number",
            "example": 94
          },
          "activeStreak": {
            "type": "number",
            "example": 19
          },
          "completedTasksCount": {
            "type": "number",
            "example": 42
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "Project": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "proj_1"
          },
          "name": {
            "type": "string",
            "example": "Core Platform Engine"
          },
          "key": {
            "type": "string",
            "example": "CPE"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "on_track",
              "at_risk",
              "delayed"
            ],
            "example": "on_track"
          },
          "progress": {
            "type": "number",
            "example": 78
          },
          "totalTasks": {
            "type": "number",
            "example": 28
          },
          "completedTasks": {
            "type": "number",
            "example": 22
          },
          "repoUrl": {
            "type": "string"
          },
          "leadId": {
            "type": "string",
            "example": "usr_1"
          },
          "lead": {
            "$ref": "#/components/schemas/User"
          },
          "team": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/User"
            }
          },
          "deadline": {
            "type": "string",
            "example": "2026-09-25"
          },
          "color": {
            "type": "string",
            "example": "#6366f1"
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "UpdateProjectRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "key": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "on_track",
              "at_risk",
              "delayed"
            ]
          },
          "repoUrl": {
            "type": "string"
          },
          "leadId": {
            "type": "string"
          },
          "teamIds": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "deadline": {
            "type": "string"
          },
          "color": {
            "type": "string"
          }
        }
      },
      "ProjectResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Project"
          }
        }
      },
...
      "Task": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "task_1"
          },
          "key": {
            "type": "string",
            "example": "CPE-104"
          },
          "title": {
            "type": "string",
            "example": "Implement Redis Bloom filter for cache hit optimization"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "backlog",
              "in_progress",
              "in_review",
              "done"
            ],
            "example": "in_progress"
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ],
            "example": "urgent"
          },
          "projectId": {
            "type": "string",
            "example": "proj_1"
          },
          "projectName": {
            "type": "string",
            "example": "Core Platform Engine"
          },
          "assigneeId": {
            "type": "string",
            "example": "usr_1"
          },
          "assignee": {
            "$ref": "#/components/schemas/User"
          },
          "storyPoints": {
            "type": "number",
            "example": 5
          },
          "dueDate": {
            "type": "string",
            "example": "2026-09-08"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "example": [
              "Backend",
              "Redis"
            ]
          },
          "createdAt": {
            "type": "string"
          },
          "updatedAt": {
            "type": "string"
          }
        }
      },
...
      "UpdateTaskRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "backlog",
              "in_progress",
              "in_review",
              "done"
            ]
          },
          "priority": {
            "type": "string",
            "enum": [
              "low",
              "medium",
              "high",
              "urgent"
            ]
          },
          "projectId": {
            "type": "string"
          },
          "assigneeId": {
            "type": "string"
          },
          "storyPoints": {
            "type": "integer"
          },
          "dueDate": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "TaskResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "statusCode": {
            "type": "integer",
            "example": 200
          },
          "message": {
            "type": "string"
          },
          "data": {
            "$ref": "#/components/schemas/Task"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Enter JWT Bearer token obtained from /api/auth/login or /api/auth/register"
      }
    }
  }
}